git-svn-id: https://subversion.assembla.com/svn/smart-cs/trunk@5609 37136b1e-4cb1-4992-8394-573b48a9bdce

+++ source/java/org.wcs.smart.patrol/src/org/wcs/smart/patrol/internal/ui/PatrolLegsComposite.java
@@ -33,6 +33,8 @@ import java.util.HashSet;
 import java.util.Iterator;
 import java.util.List;
 
+import org.eclipse.core.runtime.IConfigurationElement;
+import org.eclipse.core.runtime.Platform;
 import org.eclipse.jface.dialogs.MessageDialog;
 import org.eclipse.jface.viewers.ISelectionChangedListener;
 import org.eclipse.jface.viewers.IStructuredSelection;
@@ -50,7 +52,9 @@ import org.eclipse.swt.widgets.Label;
 import org.eclipse.swt.widgets.Link;
 import org.eclipse.swt.widgets.Shell;
 import org.hibernate.Session;
+import org.hsqldb.lib.HashMap;
 import org.wcs.smart.ca.Employee;
+import org.wcs.smart.patrol.IPatrolEditContribution;
 import org.wcs.smart.patrol.PatrolEventManager;
 import org.wcs.smart.patrol.PatrolHibernateManager;
 import org.wcs.smart.patrol.PatrolUtils;
@@ -71,6 +75,8 @@ import org.wcs.smart.util.SmartUtils;
  * Patrol item composite that modifies the patrol legs.  Allows users
  * to add, remove, and change patrol legs.
  * 
+ * added merge leg option - Nov 2016 - Jeff
+ * 
  * @author Emily
  * @since 1.0.0
  */
@@ -82,6 +88,7 @@ public class PatrolLegsComposite extends PatrolItemComposite{
 	private static final String END_INFO_LABEL = Messages.PatrolLegsComposite_PatrolEnd_Label;
 	
 	private Label lblDateInfo;
+	private Label lblNewPatrol;
 	private PatrolLegTable patrolLegViewer;
 	private Patrol patrol;
 	
@@ -99,6 +106,8 @@ public class PatrolLegsComposite extends PatrolItemComposite{
 	private Link lnkEditDate;
 	private Composite main;
 	
+	private ArrayList<Patrol> newPatrols = new ArrayList<Patrol>();
+	private HashMap movedPoints = new HashMap();
 
 	/**
 	 * Creates a new patrol legs composite
@@ -137,7 +146,7 @@ public class PatrolLegsComposite extends PatrolItemComposite{
 		main = new Composite(parent, SWT.NONE);
 		main.setLayoutData(new GridData(SWT.FILL, SWT.FILL, true, true));
 		main.setLayout(new GridLayout(1, false));
-		
+
 		if (canEditDates){
 			Composite tmp = new Composite(main, SWT.NONE);
 			tmp.setLayoutData(new GridData(SWT.LEFT, SWT.FILL, false, false));
@@ -166,9 +175,10 @@ public class PatrolLegsComposite extends PatrolItemComposite{
 		
 		patrolLegViewer = new PatrolLegTable();
 		patrolLegViewer.createTable(main);
+
 		
 		Composite buttonPanel = new Composite(main, SWT.NONE);
-		buttonPanel.setLayout(new GridLayout(5, false));
+		buttonPanel.setLayout(new GridLayout(7, false));
 		
 		final Button btnChangeTransport = new Button(buttonPanel, SWT.PUSH);
 		btnChangeTransport.setText(Messages.PatrolLegsComposite_ChangeTransport_Button);
@@ -258,16 +268,73 @@ public class PatrolLegsComposite extends PatrolItemComposite{
 				}
 			}
 		});
+
+		
+		final Button btnmergeLegs = new Button(buttonPanel, SWT.PUSH);
+		btnmergeLegs.setText(Messages.PatrolLegsComposite_MergeLegs);
+		btnmergeLegs.addSelectionListener(new SelectionAdapter() {
+			@Override
+			public void widgetSelected(SelectionEvent e){
+				IStructuredSelection selection = (IStructuredSelection)patrolLegViewer.getSelection();
+				ArrayList<PatrolLeg> toBeMerged = new ArrayList<PatrolLeg>(); 
+				Iterator<?> i = selection.iterator();
+			    while ( i.hasNext() ) {
+			    	toBeMerged.add( ((PatrolLeg)i.next()) );
+			    }
+				MergePatrolLegDialog patrolLegDialog = new MergePatrolLegDialog(getShell(), toBeMerged, typeOps, session);
+				if (patrolLegDialog.open() == Window.OK){
+					patrolLegDialog.getNewLeg().setPatrol(patrol);
+					legs.add(patrolLegDialog.getNewLeg());
+					for(PatrolLeg pld : toBeMerged){
+						legs.remove(pld);
+					}
+					sortAndRefresh();
+					fireChangeListeners();
+				}
+			}
+		});
+
+		
+		final Button btnMoveToNewPatrol = new Button(buttonPanel, SWT.PUSH);
+		btnMoveToNewPatrol.setText(Messages.PatrolLegsComposite_0);
+		btnMoveToNewPatrol.addSelectionListener(new SelectionAdapter() {
+			@Override
+			public void widgetSelected(SelectionEvent e){
+				IStructuredSelection selection = (IStructuredSelection)patrolLegViewer.getSelection();
+				ArrayList<PatrolLeg> toBeMoved = new ArrayList<PatrolLeg>(); 
+				Iterator<?> i = selection.iterator();
+			    while ( i.hasNext() ) {
+			    	toBeMoved.add( ((PatrolLeg)i.next()) );
+			    }
+			    MovePatrolLegDialog movePatrolLegDialog = new MovePatrolLegDialog(getShell(), toBeMoved, session);
+				if (movePatrolLegDialog.open() == Window.OK){
+					for(PatrolLeg pld : toBeMoved){
+						legs.remove(pld);
+					}
+					sortAndRefresh();
+					fireChangeListeners();
+					//add the new patrol of a list of new ones so we can save it later (users can use the option multiple times beforing saving, we want to save them all).
+					if(movePatrolLegDialog.getNewPatrol() != null) newPatrols.add(movePatrolLegDialog.getNewPatrol());
+					lblNewPatrol.setText(Messages.PatrolLegsComposite_1);
+				}
+			}
+		});
+
+
+		lblNewPatrol = new Label(main, SWT.NONE);
+		lblNewPatrol.setLayoutData(new GridData(SWT.FILL, SWT.FILL, true, false));
 		
 		patrolLegViewer.getTable().addSelectionChangedListener(new ISelectionChangedListener() {			
 			@Override
 			public void selectionChanged(SelectionChangedEvent event) {
-				boolean isEmpty =  ((IStructuredSelection)patrolLegViewer.getSelection()).isEmpty();
-				btnChangeTransport.setEnabled(!isEmpty);
-				btnAddLeg.setEnabled(!isEmpty);
-				btnSplit.setEnabled(!isEmpty);
-				btnRemoveLeg.setEnabled(!isEmpty && legs.size() > 1);
-				btnEditLeg.setEnabled(!isEmpty);
+				int numSelected =  ((IStructuredSelection)patrolLegViewer.getSelection()).size();
+				btnChangeTransport.setEnabled(numSelected == 1);
+				btnAddLeg.setEnabled(numSelected == 1);
+				btnSplit.setEnabled(numSelected == 1);
+				btnRemoveLeg.setEnabled((numSelected == 1) && legs.size() > 1);
+				btnEditLeg.setEnabled(numSelected == 1);
+				btnmergeLegs.setEnabled(numSelected > 1);
+				btnMoveToNewPatrol.setEnabled(numSelected > 0 && legs.size() > 1 && numSelected != legs.size());
 			}
 		});
 		
@@ -276,6 +343,8 @@ public class PatrolLegsComposite extends PatrolItemComposite{
 		btnRemoveLeg.setEnabled(false );
 		btnEditLeg.setEnabled( false );
 		btnChangeTransport.setEnabled(false);
+		btnmergeLegs.setEnabled(false);
+		btnMoveToNewPatrol.setEnabled(false);
 		
 		return main;
 	}
@@ -375,22 +444,19 @@ public class PatrolLegsComposite extends PatrolItemComposite{
 	}
 
 	private PatrolLeg clonePatrolLeg(PatrolLeg leg){
-		PatrolLeg tmpLeg = new PatrolLeg();
-		tmpLeg.setPatrol(leg.getPatrol());
-		tmpLeg.setId(leg.getId());
+		PatrolLeg tmpLeg = leg.simpleClone();
 		
-		//start time
-		tmpLeg.setStartDate(leg.getStartDate());
-		tmpLeg.setEndDate(leg.getEndDate());
-		//type
-		tmpLeg.setType(leg.getType());
-		//members
-		tmpLeg.setMembers(new ArrayList<PatrolLegMember>());
-		for (PatrolLegMember mem : leg.getMembers()){
-			PatrolLegMember clone = mem.clone();
-			clone.setPatrolLeg(tmpLeg);
-			tmpLeg.getMembers().add(clone);
+		tmpLeg.setPatrolLegDays(new ArrayList<PatrolLegDay>());
+		if (leg.getPatrolLegDays() != null && leg.getPatrolLegDays().size() > 0){
+			//Clone Leg Days as well, now that we can merge we need more details on these.
+			for (PatrolLegDay pld : leg.getPatrolLegDays()){
+				PatrolLegDay clone = pld.clone();
+				clone.setWaypoints(pld.getWaypoints()); //This is kinda bad as the waypoints are now associated with both clone and pld, but the only times 'clone' is saved we create new PatrolWaypoints objects, and the the pld is then deleted, so I think that makes it OK, maybe. Blame Jeff if not.
+				clone.setPatrolLeg(tmpLeg);
+				tmpLeg.getPatrolLegDays().add(clone);
+			}
 		}
+		
 		//uuid
 		tmpLeg.setUuid(leg.getUuid());
 		return tmpLeg;
@@ -417,6 +483,23 @@ public class PatrolLegsComposite extends PatrolItemComposite{
 		ArrayList<PatrolLeg> allLegs = new ArrayList<PatrolLeg>();
 		allLegs.addAll(legs);
 		session.flush();
+		
+		
+		//if the user used the "split into a new patrol" option, don't delete the points we are going to move
+		session.flush();
+		for(Patrol p2 : newPatrols){
+			//put all the waypoints into our 'moved' hash so they don't get deleted when we delete the leg from the old patrol
+			for(PatrolLeg pl : p2.getLegs()){
+				for(PatrolLegDay pld : pl.getPatrolLegDays()){
+					for(PatrolWaypoint pwp : pld.getWaypoints()){
+						movedPoints.put(pwp.getWaypoint().getUuid(), "true");	 //$NON-NLS-1$
+					}
+				}
+			}
+		}
+		session.flush();
+		
+		
 		for (Iterator<PatrolLeg> iterator = allLegs.iterator(); iterator.hasNext();) {
 			PatrolLeg updatedLeg = (PatrolLeg) iterator.next();
 			if (updatedLeg.getUuid() != null){
@@ -453,13 +536,23 @@ public class PatrolLegsComposite extends PatrolItemComposite{
 		}
 		
 		//new legs
+  
 		for (Iterator<PatrolLeg> iterator = allLegs.iterator(); iterator.hasNext();) {
 			PatrolLeg newLeg = (PatrolLeg) iterator.next();
 			
 			PatrolLeg clone = clonePatrolLeg(newLeg);
 			clone.setPatrol(p);
 			p.getLegs().add(clone);
+			//save the waypoints in each leg-day, they are not cascaded like everything else, presumably for a reason, so I don't want to break everything else by changing it.
+			for(PatrolLegDay pld : clone.getPatrolLegDays()){
+				if (pld.getWaypoints() != null){
+				for(PatrolWaypoint wp : pld.getWaypoints()){
+					wp.setPatrolLegDay(pld);
+					movedPoints.put(wp.getWaypoint().getUuid(), "true"); //$NON-NLS-1$
+				}}
+			}
 		}
+		session.flush();
 		
 		//legs no longer used; these must be removed
 		for (PatrolLeg toRemove: currentLegs){
@@ -468,7 +561,9 @@ public class PatrolLegsComposite extends PatrolItemComposite{
 				for (PatrolLegDay pld : toRemove.getPatrolLegDays()){
 					if (pld.getWaypoints() != null){
 						for (PatrolWaypoint pw : pld.getWaypoints()){
-							session.delete(pw.getWaypoint());
+							if(movedPoints.get(pw.getWaypoint().getUuid()) == null){ //only delete the waypoints if they didn't get moved into a new leg.
+								session.delete(pw.getWaypoint());
+							}
 						}
 					}
 				}
@@ -481,6 +576,47 @@ public class PatrolLegsComposite extends PatrolItemComposite{
 		p.createLegDays(session);
 		p.recalculateType();
 		session.saveOrUpdate(p);
+	
+		
+		for(PatrolLeg l : p.getLegs()){
+			for(PatrolLegDay d : l.getPatrolLegDays()){
+				if(d.getWaypoints() != null){
+					for(PatrolWaypoint w : d.getWaypoints()){
+						session.saveOrUpdate(w);
+					}
+				}
+			}
+		}
+		session.flush();
+		
+		
+		//if we made brand new patrols, save them and copy any plan and intel links there were
+		for(Patrol p2 : newPatrols){
+			session.saveOrUpdate(p2);
+			//save the waypoints, they are not cascaded like everything else.
+			for(PatrolLeg pl : p2.getLegs()){
+				for(PatrolLegDay pld : pl.getPatrolLegDays()){
+					for(PatrolWaypoint pwp : pld.getWaypoints()){
+						//pwp.setPatrolLegDay(pld);
+						session.saveOrUpdate(pwp);
+					}
+				}
+			}
+			
+			//copy the intel and plan links to the new patrol as well
+			//do this by running the contribution from each plug-in that has links to patrols (intel and plan at the time of writing)
+			List<IPatrolEditContribution> contributions = findContributions();
+			for(IPatrolEditContribution c : contributions){
+				c.splitPatrol(session, patrol, p2);
+			}
+			
+			
+			lblNewPatrol.setText(Messages.PatrolLegsComposite_2);
+		}
+		
+
+		
+		session.flush();
 		return true;
 			
 	}
@@ -584,6 +720,26 @@ public class PatrolLegsComposite extends PatrolItemComposite{
 		}		
 		return null;		
 	}
+	
+	private List<IPatrolEditContribution> findContributions(){
+
+		List<IPatrolEditContribution> items = new ArrayList<IPatrolEditContribution>();
+		if (Platform.getExtensionRegistry() == null) return
+		Collections.emptyList();
+		IConfigurationElement[] config = Platform.getExtensionRegistry().getConfigurationElementsFor(IPatrolEditContribution.EXTENSION_ID);
+		try {
+		    for (IConfigurationElement e : config) {
+		        if (e.getName().equals("edit")){ //$NON-NLS-1$
+		            IPatrolEditContribution page = (IPatrolEditContribution)e.createExecutableExtension("class"); //$NON-NLS-1$
+		            items.add(page);
+		        }
+		    }
+		}catch (Exception ex){
+		         SmartPatrolPlugIn.displayLog(Messages.CreatePatrolWizard_ErrorCreatingWizardPages, ex);
+		    return null;
+		}
+		return items;
+	} 
 	
 	/**
 	 * @see org.wcs.smart.patrol.internal.ui.PatrolItemComposite#getAttribute()
